microservices/requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
kafka_python==2.0.2
pandas==1.5.1
pandas_profiling==3.6.6
psutil==5.9.3
pymongo==4.3.3
Requests==2.31.0
scikit_learn==1.2.0
ydata_profiling==4.5.1